code/requirements.txt
google-cloud-speech==3.0.2
google-cloud-storage==2.10.0
pydub==0.25.1
